Port to GTK 3
Unfortunately, the move to GTK 3 makes the main (study) window wider
even without an image loaded (#901677).  This is because some of the
widgets are larger, like GtkSpinButton with big "+" and "-" in it.
Additionally, GTK computes the width of the widget based on the
min/max values of the adjustment -- so the "thickness" button gets
wider than usual because the maximum allowed value is G_MAXDOUBLE.

A major hurdle was frequent aborts when using the custom log handler
due to a GLib bug (see src/amide.h).  The only solution I came up
with was to redefine the macros and avoid setting a custom handler.

The GtkCurve and GtkGammaCurve widgets were ported to GTK 3 but
stripped from some functionality that AMIDE doesn't need (functions
like gtk_curve_set_vector are not available, also the additional
buttons of the GtkGammaCurve widget).  As is required by GTK 3,
drawing is done to an off-screen buffer (a cairo surface) which is
then effectively drawn on the widget during the ::draw signal.

The GtkHandleBox widget was not ported so parts of the UI are no
longer detachable.  I didn't consider this functionality important
enough to bother (unlike GtkCurve/GtkGammaCurve).

The AmitkDial widget was modified to behave as with policy
GTK_UPDATE_DISCONTINUOUS, which is the only use within AMIDE.  Since
GTK 3 removed policy support in GtkRange-based widgets, all of the
widgets where GTK_UPDATE_DISCONTINUOUS is set (for a good reason)
were modified to connect to a "button-release-event" handler, as a
workaround.  The "value-changed" callback is invoked from there.

Fixes: #27