Port to GSettings
There is no migration code; users will lose all their configuration.

I don't know why GConf was not used on Windows and macOS; I guess
it's because it is a different library.  Someone (upstream) with
access to these systems should verify if GSettings work there (it's
part of GLib so it should in theory) -- if that's the case it would
simplify the code a lot as much of amide_gconf.c will be removed
along with the *_with_default functions.

Addresses: #27