Make the main (study) window resizable
Toolbar menu items "Thickness", "Field of view" and "Zoom" do nothing
but they are necessary -- if the window is resized so that the right
border is at one of these widgets, no drop-down arrow will be shown
so the user will not have access to the Gate/Time buttons.  Other
items from "Threshold" leftwards are proper GtkToolItems (rather than
widgets packed in a GtkToolItem) so their menus are fully functional
and will appear as normal/toggle/radio menu items.

Fixes #26.